% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uisegment.R
\name{box}
\alias{box}
\title{Create a box.}
\usage{
box(..., title = "", color = "")
}
\arguments{
\item{...}{UI elements to include within the box.}

\item{title}{Label of the box.}

\item{color}{Color of the sidebar. One of \code{c("", "red", "orange", "yellow", "olive", "green", "teal", "blue", "violet", "purple", "pink", "brown", "grey", "black")}}
}
\value{
A box that can be passed to \code{\link[semantic.dashboard]{dashboardBody}}
}
\description{
Create a box with additional UI elements.
}
\examples{
if(interactive()){

  library(shiny)
  library(semantic.dashboard)

  ui <- dashboardPage(
    dashboardHeader(color = "blue"),
    dashboardSidebar(side = "top", size = "thin", color = "teal",
                     menuItem("tab1", "Tab 1"),
                     menuItem("tab2", "Tab 2")),
    dashboardBody(tabItem(tabName = "tab1", active = TRUE, p("Tab 1")),
                  tabItem(tabName = "tab2", p("Tab 2")))
  )

  server <- function(input, output) {
  }

  shinyApp(ui, server)
}
}
