% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column.R
\name{column}
\alias{column}
\title{Create a column.}
\usage{
column(width, ...)
}
\arguments{
\item{width}{Width of the column. Between 1 and 16.}

\item{...}{UI elements to include within the column.}
}
\value{
A column that can be passed to \code{\link[semantic.dashboard]{dashboardPage}}
}
\description{
Create a column with additional UI elements.
}
