% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_dashboard.R
\name{dashboard_body}
\alias{dashboard_body}
\title{Create a body of a dashboard.}
\usage{
dashboard_body(...)
}
\arguments{
\item{...}{UI elements to include within the body.}
}
\value{
A tab that can be passed to \code{\link[semantic.dashboard]{dashboardPage}}
}
\description{
Create a body of a dashboard with tabs and other additional UI elements.
}
