% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_dashboard.R
\name{dashboard_header}
\alias{dashboard_header}
\alias{dashboardHeader}
\title{Create a header of a dashboard.}
\usage{
dashboard_header(..., color = "black")

dashboardHeader(..., color = "black")
}
\arguments{
\item{...}{UI elements to include within the header.}

\item{color}{Color of the sidebar. One of \code{c("red", "orange", "yellow", "olive", "green", "teal", "blue", "violet", "purple", "pink", "brown", "grey", "black")}}
}
\value{
A header that can be passed to \code{\link[semantic.dashboard]{dashboardPage}}
}
\description{
Create a header of a dashboard with other additional UI elements.
}
\section{Functions}{
\itemize{
\item \code{dashboardHeader}: Create a header of a dashboard.
}}

\examples{
if(interactive()){

  library(shiny)
  library(semantic.dashboard)

  ui <- dashboardPage(
    dashboardHeader(color = "blue"),
    dashboardSidebar(side = "top", size = "thin", color = "teal",
                     menuItem("tab1", "Tab 1"),
                     menuItem("tab2", "Tab 2")),
    dashboardBody(tabItem(tabName = "tab1", active = TRUE, p("Tab 1")),
                  tabItem(tabName = "tab2", p("Tab 2")))
  )

  server <- function(input, output) {
  }

  shinyApp(ui, server)
}
}
