% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_dashboard.R
\name{dashboard_page}
\alias{dashboard_page}
\alias{dashboardPage}
\title{Create a dashboard.}
\usage{
dashboard_page(dashboardHeader, dashboardSidebar, dashboardBody, title = "",
  suppress_bootstrap = TRUE)

dashboardPage(dashboardHeader, dashboardSidebar, dashboardBody, title = "",
  suppress_bootstrap = TRUE)
}
\arguments{
\item{dashboardHeader}{Header of a dashboard.}

\item{dashboardSidebar}{Sidebar of a dashboard.}

\item{dashboardBody}{Body of a dashboard.}

\item{title}{Title of a dashboard.}

\item{suppress_bootstrap}{There are some conflicts in CSS styles between SemanticUI and Bootstrap. For the time being it's better to suppress Bootstrap. If \code{TRUE} bootstrap dependency from \code{shiny} will be disabled.}
}
\value{
Dashboard.
}
\description{
Create a page with menu item sidebar and body containing tabs and other additional elements.
}
\section{Functions}{
\itemize{
\item \code{dashboardPage}: Create a dashboard.
}}

\examples{
if(interactive()){

  library(shiny)
  library(semantic.dashboard)

  ui <- dashboardPage(
    dashboardSidebar(side = "top", size = "thin", color = "teal",
                     menuItem("tab1", "Tab 1"),
                     menuItem("tab2", "Tab 2")),
    dashboardBody(tabItem(tabName = "tab1", active = TRUE, p("Tab 1")),
                  tabItem(tabName = "tab2", p("Tab 2")))
  )

  server <- function(input, output) {
  }

  shinyApp(ui, server)
}
}
