% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_dashboard.R
\name{dashboardBody}
\alias{dashboardBody}
\title{Create a body of a dashboard.}
\usage{
dashboardBody(...)
}
\arguments{
\item{...}{UI elements to include within the pusher.}
}
\value{
A tab that can be passed to \code{semanticDashboard}
}
\description{
Create a body of a dashboard.
}
\examples{
dashboardDoby(uitab(id = "a", active = TRUE,
  numericInput("n1", "n", 50),
  plotOutput("plot1"),
  numericInput("n2", "n", 50),
  plotOutput("plot2"),
  numericInput("n3", "n", 50),
  plotOutput("plot3")),
  uitab(id = "b",
        numericInput("n4", "nn", 50),
        plotOutput("plot4")),
  uitab(id = "c",
        numericInput("n5", "nnn", 50),
        plotOutput("plot5")))
}
