% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdown_menu.R
\name{dropdown_menu}
\alias{dropdown_menu}
\alias{dropdownMenu}
\title{Create a dropdown menu (alias for \code{dropdown_menu} for compatibility with \code{shinydashboard})}
\usage{
dropdown_menu(..., type = "messages")

dropdownMenu(..., type = "messages")
}
\arguments{
\item{...}{UI elements to include within the dropdown menu.}

\item{type}{Type of the displayed items.}
}
\value{
A dropdown menu that can be passed to \code{\link[semantic.dashboard]{dashboardHeader}}
}
\description{
Create a dropdown menu with additional UI elements.
}
\section{Functions}{
\itemize{
\item \code{dropdownMenu}: Create a dropdown menu (alias for \code{dropdown_menu} for compatibility with \code{shinydashboard})
}}

