% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdown_menu_output.R
\name{dropdown_menu_output}
\alias{dropdown_menu_output}
\alias{dropdownMenuOutput}
\title{Create a dropdown menu output (alias for \code{dropdown_menu_output} for compatibility with \code{shinydashboard})}
\usage{
dropdown_menu_output(outputId)

dropdownMenuOutput(outputId)
}
\arguments{
\item{outputId}{Id of the output.}
}
\value{
A dropdown menu that can be passed to \code{\link[semantic.dashboard]{dashboardHeader}}
}
\description{
UI-side function for dynamic valueBox.
}
\section{Functions}{
\itemize{
\item \code{dropdownMenuOutput}: Create a value box output (alias for \code{dropdown_menu_output} for compatibility with \code{shinydashboard})
}}

