% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{get_dashboard_dependencies}
\alias{get_dashboard_dependencies}
\title{Get the semantic.dashboard dependencies}
\usage{
get_dashboard_dependencies()
}
\value{
semantic.dashboard dependencies
}
\description{
To add dependencies in the future follow the \code{\link[htmltools]{htmlDependency}} help.
}
