% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_dashboard.R
\name{get_dependencies}
\alias{get_dependencies}
\title{Add dashboard dependencies to html}
\usage{
get_dependencies()
}
\value{
Content with appended dependencies.
}
\description{
Internal function that adds dashboard dependencies to html.
}
