% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uimenu_item.R
\name{icon}
\alias{icon}
\title{Create Semantic UI icon tag}
\usage{
icon(type = "", ...)
}
\arguments{
\item{type}{A name of an icon. Look at http://semantic-ui.com/elements/icon.html for all possibilities.}

\item{...}{Other arguments to be added as attributes of the tag (e.g. style, class etc.)}
}
\description{
This creates an icon tag using Semantic UI styles.
}
