% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uistatistic.R
\name{infoBox}
\alias{infoBox}
\title{Create a infoBox.}
\usage{
infoBox(title, value, icon = NULL, color = "blue")
}
\arguments{
\item{title}{Label of the infoBox.}

\item{value}{Value of the infoBox.}

\item{icon}{Icon of the infoBox.}

\item{color}{Color of the box. One of \code{c("", "red", "orange", "yellow", "olive", "green", "teal", "blue", "violet", "purple", "pink", "brown", "grey", "black")}}
}
\value{
A box that can be passed to \code{\link[semantic.dashboard]{dashboardBody}}
}
\description{
Create a infoBox with additional UI elements.
}
\examples{
if(interactive()){

  library(shiny)
  library(semantic.dashboard)

  ui <- dashboardPage(
    dashboardHeader(color = "blue"),
    dashboardSidebar(side = "top", size = "thin", color = "teal",
                     menuItem("tab1", "Tab 1"),
                     menuItem("tab2", "Tab 2")),
    dashboardBody(tabItem(tabName = "tab1", active = TRUE, p("Tab 1")),
                  tabItem(tabName = "tab2", p("Tab 2")))
  )

  server <- function(input, output) {
  }

  shinyApp(ui, server)
}
}
