% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{light_semantic_palette}
\alias{light_semantic_palette}
\title{Semantic light colors
https://github.com/Semantic-Org/Semantic-UI/blob/master/src/themes/default/globals/site.variables}
\format{
An object of class \code{character} of length 13.
}
\usage{
light_semantic_palette
}
\description{
Semantic light colors
https://github.com/Semantic-Org/Semantic-UI/blob/master/src/themes/default/globals/site.variables
}
\keyword{datasets}
