% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uimenu_item.R
\name{menu_item}
\alias{menu_item}
\title{Create a menu item.}
\usage{
menu_item(tabName, label, icon = NULL)
}
\arguments{
\item{tabName}{Id of the tab.}

\item{label}{Label of the menu item.}

\item{icon}{Icon of the menu item. (Optional)}
}
\value{
A menu item that can be passed \code{\link[semantic.dashboard]{dashboardSidebar}}
}
\description{
Create a menu item corresponding to a tab.
}
