% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar_menu_output.R
\name{menu_item_output}
\alias{menu_item_output}
\alias{menuItemOutput}
\title{Create a menu item output.}
\usage{
menu_item_output(outputId)

menuItemOutput(outputId)
}
\arguments{
\item{outputId}{Id of the output.}
}
\value{
A menu item that can be passed to \code{\link[semantic.dashboard]{sidebarMenu}}
}
\description{
UI-side function for dynamic manuItem.
}
\section{Functions}{
\itemize{
\item \code{menuItemOutput}: Create a menu item output (alias for \code{menu_item_output} for compatibility with \code{shinydashboard})
}}

