% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uimenu_item.R
\name{menuItem}
\alias{menuItem}
\title{Create a menu item.}
\usage{
menuItem(tabName, label, icon = NULL)
}
\arguments{
\item{tabName}{Id of the tab.}

\item{label}{Label of the menu item.}

\item{icon}{Icon of the menu item. (Optional)}
}
\value{
A menu item that can be passed \code{\link[semantic.dashboard]{dashboardSidebar}}
}
\description{
Create a menu item corresponding to a tab.
}
\examples{
if(interactive()){

  library(shiny)
  library(semantic.dashboard)

  ui <- dashboardPage(
    dashboardHeader(color = "blue"),
    dashboardSidebar(side = "top", size = "thin", color = "teal",
                     menuItem("tab1", "Tab 1"),
                     menuItem("tab2", "Tab 2")),
    dashboardBody(tabItem(tabName = "tab1", active = TRUE, p("Tab 1")),
                  tabItem(tabName = "tab2", p("Tab 2")))
  )

  server <- function(input, output) {
  }

  shinyApp(ui, server)
}
}
