% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uidropdown_menu.R
\name{message_item}
\alias{message_item}
\alias{messageItem}
\title{Create a message item (alias for \code{message_item} for compatibility with \code{shinydashboard})}
\usage{
message_item(from, message, icon = "user", color = "")

messageItem(from, message, icon = "user", color = "")
}
\arguments{
\item{from}{Who the message is from.}

\item{message}{Text of the message.}

\item{icon}{Additional icon.}

\item{color}{Color of the message item. One of \code{c("", "red", "orange", "yellow", "olive", "green", "teal", "blue", "violet", "purple", "pink", "brown", "grey", "black")}}
}
\value{
A message item that can be passed to \code{\link[semantic.dashboard]{dropdownMenu}}
}
\description{
Create a message item.
}
\section{Functions}{
\itemize{
\item \code{messageItem}: Create a message item (alias for \code{message_item} for compatibility with \code{shinydashboard})
}}

