% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uicolumn.R
\name{num_to_string}
\alias{num_to_string}
\title{Create a column.}
\usage{
num_to_string(number)
}
\arguments{
\item{number}{Number to convert.}
}
\value{
A string corresponding to inputed number.
}
\description{
Transforms number to corresponding string.
}
