% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar_menu_output.R
\name{render_menu}
\alias{render_menu}
\alias{renderMenu}
\title{Create a menu output.}
\usage{
render_menu(expr, env = parent.frame(), quoted = FALSE)

renderMenu(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{menu.}

\item{env}{The environment in which to evaluate expr.}

\item{quoted}{Is expr a quoted expression (with \code{quote()})? This is useful if you want to save an expression in a variable.}
}
\value{
A dynamic menu that can be assigned to output.
}
\description{
Server-side function for dynamic sidebarMenu.
}
\section{Functions}{
\itemize{
\item \code{renderMenu}: Create a menu output (alias for \code{render_menu} for compatibility with \code{shinydashboard})
}}

