% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uivalue_box_output.R
\name{render_value_box}
\alias{render_value_box}
\title{Create a valueBox output.}
\usage{
render_value_box(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{ValueBox.}

\item{env}{The environment in which to evaluate expr.}

\item{quoted}{Is expr a quoted expression (with \code{quote()})? This is useful if you want to save an expression in a variable.}
}
\value{
A dynamic valueBox that can be passed to \code{\link[semantic.dashboard]{valueBoxOutput}}
}
\description{
Server-side function for dynamic valueBox.
}
