% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar_menu_output.R
\name{sidebar_menu_output}
\alias{sidebar_menu_output}
\alias{sidebarMenuOutput}
\title{Create a sidebar menu output.}
\usage{
sidebar_menu_output(outputId)

sidebarMenuOutput(outputId)
}
\arguments{
\item{outputId}{Id of the output.}
}
\value{
A sidebar menu that can be passed to \code{\link[semantic.dashboard]{dashboardSidebar}}
}
\description{
UI-side function for dynamic sidebarMenu.
}
\section{Functions}{
\itemize{
\item \code{sidebarMenuOutput}: Create a sidebar menu output (alias for \code{sidebar_menu_output} for compatibility with \code{shinydashboard})
}}

