% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab_item}
\alias{tab_item}
\title{Create a tab panel (alias for \code{tab_item} for compatibility with \code{shinydashboard})}
\usage{
tab_item(tabName, ...)
}
\arguments{
\item{tabName}{Id of the tab.}

\item{...}{UI elements to include within the tab.}
}
\value{
A tab that can be passed to \code{\link[semantic.dashboard]{dashboardBody}}
}
\description{
Create a tab panel with additional UI elements.
}
