% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab_items}
\alias{tab_items}
\title{Create a panel with tabs (alias for \code{tab_items} for compatibility with \code{shinydashboard})}
\usage{
tab_items(..., selected = 1)
}
\arguments{
\item{...}{Tabs.}

\item{selected}{Which tab should be active on start.}
}
\value{
A panel with tabs that can be passed to \code{\link[semantic.dashboard]{dashboardBody}}
}
\description{
Create a panel with tabs.
}
