% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uitab.R
\name{tabItem}
\alias{tabItem}
\title{Create a tab panel.}
\usage{
tabItem(tabName, ..., active = FALSE)
}
\arguments{
\item{tabName}{Id of the tab.}

\item{...}{UI elements to include within the tab.}

\item{active}{Should it be active.}
}
\value{
A tab that can be passed to \code{\link[semantic.dashboard]{dashboardBody}}
}
\description{
Create a tab panel with additional UI elements.
}
\examples{
if(interactive()){

  library(shiny)
  library(semantic.dashboard)

  ui <- dashboardPage(
    dashboardHeader(color = "blue"),
    dashboardSidebar(side = "top", size = "thin", color = "teal",
                     menuItem("tab1", "Tab 1"),
                     menuItem("tab2", "Tab 2")),
    dashboardBody(tabItem(tabName = "tab1", active = TRUE, p("Tab 1")),
                  tabItem(tabName = "tab2", p("Tab 2")))
  )

  server <- function(input, output) {
  }

  shinyApp(ui, server)
}
}
