% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uitab.R
\name{tabItems}
\alias{tabItems}
\title{Create a panel with tabs.}
\usage{
tabItems(..., selected = 1)
}
\arguments{
\item{...}{Tabs.}

\item{selected}{Which tab should be active on start.}
}
\value{
A panel with tabs that can be passed to \code{\link[semantic.dashboard]{dashboardBody}}
}
\description{
Create a panel with tabs.
}
