% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdown_menu.R
\name{task_item}
\alias{task_item}
\alias{taskItem}
\title{Create a task item.}
\usage{
task_item(text, value, color = "")

taskItem(text, value, color = "")
}
\arguments{
\item{text}{Progress bar label.}

\item{value}{Progress bar value.}

\item{color}{Color of the task item. One of \code{c("", "red", "orange",
"yellow", "olive", "green", "teal", "blue", "violet", "purple", "pink",
"brown", "grey", "black")}}
}
\value{
A task item that can be passed to \code{\link[semantic.dashboard]{dropdownMenu}}
}
\description{
Create a task item.
}
\section{Functions}{
\itemize{
\item \code{taskItem}: Create a task item (alias for \code{taks_item} for compatibility with \code{shinydashboard})
}}

\examples{
taskItem("Project progress...", 50.777, color = "red")
}
