% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu_item.R
\name{uimenu_item}
\alias{uimenu_item}
\title{Create a menu item.}
\usage{
uimenu_item(id, label)
}
\arguments{
\item{id}{Id of the tab.}

\item{label}{Label of the menu item.}
}
\value{
A menu item that can be passed \code{\link[semantic.dashboard]{dashboardSidebar}}
}
\description{
Create a menu item corresponding to a tab.
}
\examples{
if(interactive()){

  library(shiny)
  library(semantic.dashboard)
  library(plotly)
  library(DT)

  ui <- dashboardPage(
    dashboardSidebar(side = "top", size = "thin", color = "teal",
                     uimenu_item("plot_tab", "My plot"),
                     uimenu_item("table_tab", "My table")),
    dashboardBody(uitab(id = "plot_tab", active = TRUE,
                        selectInput(inputId =  "variable1", choices = names(mtcars),
                                    label = "Select first variable", selected = "mpg"),
                        selectInput(inputId =  "variable2", choices = names(mtcars),
                                    label = "Select second variable", selected = "cyl"),
                        plotlyOutput("mtcars_plot")),
                  uitab(id = "table_tab",
                        dataTableOutput("mtcars_table")))
  )

  server <- function(input, output) {

    output$mtcars_plot <- renderPlotly(plot_ly(mtcars, x=~mtcars[ , input$variable1], y=~mtcars[ , input$variable2], width = "100\%"))
    output$mtcars_table <- renderDataTable(mtcars)

    lapply(c("mtcars_plot", "mtcars_table"),
           function(x) outputOptions(output, x, suspendWhenHidden = FALSE))
  }

  shinyApp(ui, server)
}
}
