% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uisegment.R
\name{uisegment}
\alias{uisegment}
\title{Create a box.}
\usage{
uisegment(..., title = "", color = "")
}
\arguments{
\item{...}{UI elements to include within the box.}

\item{title}{Label of the box.}

\item{color}{Color of the sidebar. One of \code{c("", "red", "orange", "yellow", "olive", "green", "teal", "blue", "violet", "purple", "pink", "brown", "grey", "black")}}
}
\value{
A box that can be passed to \code{\link[semantic.dashboard]{dashboardBody}}
}
\description{
Create a box with additional UI elements.
}
