% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uistatistic.R
\name{uistatistic}
\alias{uistatistic}
\title{Create a valueBox.}
\usage{
uistatistic(title, value, icon = NULL, color = "blue")
}
\arguments{
\item{title}{Label of the valueBox.}

\item{value}{Value of the valueBox.}

\item{icon}{Icon of the valueBox.}

\item{color}{Color of the box. One of \code{c("", "red", "orange", "yellow", "olive", "green", "teal", "blue", "violet", "purple", "pink", "brown", "grey", "black")}}
}
\value{
A box that can be passed to \code{\link[semantic.dashboard]{dashboardBody}}
}
\description{
Create a valueBox with additional UI elements.
}
