% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uistatistic.R
\name{uistatistics}
\alias{uistatistics}
\title{Create a group of stacked infoBoxes.}
\usage{
uistatistics(...)
}
\arguments{
\item{...}{infoBoxes.}
}
\value{
A group of stacked infoBoxes that can be passed to \code{\link[semantic.dashboard]{dashboardBody}}
}
\description{
Create a group of stacked infoBoxes.
}
