% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uivalue_box.R
\name{uivalue_box}
\alias{uivalue_box}
\alias{valueBox}
\title{Create a valueBox.}
\usage{
uivalue_box(title, value, icon = NULL, color = "blue")

valueBox(title, value, icon = NULL, color = "blue")
}
\arguments{
\item{title}{Label of the valueBox.}

\item{value}{Value of the valueBox.}

\item{icon}{Icon of the valueBox.}

\item{color}{Color of the valueBox. One of \code{c("", "red", "orange", "yellow", "olive", "green", "teal", "blue", "violet", "purple", "pink", "brown", "grey", "black")}}
}
\value{
A valueBox that can be passed to \code{\link[semantic.dashboard]{dashboardBody}}
}
\description{
Create a valueBox with additional UI elements.
}
\section{Functions}{
\itemize{
\item \code{valueBox}: Create a valueBox.
}}

