% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uivalue_box_output.R
\name{uivalue_box_output}
\alias{uivalue_box_output}
\title{Create a valueBox output.}
\usage{
uivalue_box_output(outputId)
}
\arguments{
\item{outputId}{Id of the output.}
}
\value{
A valueBox that can be passed to \code{\link[semantic.dashboard]{dashboardBody}}
}
\description{
UI-side function for dynamic valueBox.
}
