% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uistatistic.R
\name{uivalue_boxes}
\alias{uivalue_boxes}
\title{Create a group of stacked valueBoxes.}
\usage{
uivalue_boxes(...)
}
\arguments{
\item{...}{valueBoxes.}
}
\value{
A group of stacked valueBoxes that can be passed to \code{\link[semantic.dashboard]{dashboardBody}}
}
\description{
Create a group of stacked valueBoxes.
}
