% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu_item.R
\name{update_tab_items}
\alias{update_tab_items}
\alias{updateTabItems}
\title{Change the selected tab on the client}
\usage{
update_tab_items(session = shiny::getDefaultReactiveDomain(), tab)

updateTabItems(session = shiny::getDefaultReactiveDomain(), tab)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer}

\item{tab}{The name of the tab that should be selected}
}
\description{
Change the selected tab on the client
}
\section{Functions}{
\itemize{
\item \code{updateTabItems}: Change the selected item on the client (alias for \code{update_tab_items} for
compatibility with \code{shinydashboard})
}}

\examples{
if (interactive()) {
  ui <- dashboardPage(
    dashboardSidebar(
      sidebarMenu(
        menuItem("Tab 1", tabName = "tab1"),
        menuItem("Tab 2", tabName = "tab2")
      )
    ),
    dashboardBody(
      tabItems(
        tabItem(tabName = "tab1", h2("Tab 1")),
        tabItem(tabName = "tab2", h2("Tab 2"))
     )
   )
  )

  server <- function(input, output, session) {
    update_tab_items(tab = "tab2")
  }

  shinyApp(ui, server)
}
}
