% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu_item.R
\name{validate_tab_name}
\alias{validate_tab_name}
\title{Valid tab name should not containt dot character '.'.}
\usage{
validate_tab_name(name)
}
\arguments{
\item{name}{Tab name to validate.}
}
\description{
Valid tab name should not containt dot character '.'.
}
