% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uivalue_box.R
\name{value_box}
\alias{value_box}
\alias{valueBox}
\alias{infoBox}
\title{Create a valueBox (alias for \code{value_box} for compatibility with \code{shinydashboard})}
\usage{
value_box(title, value, icon = NULL, color = "blue", width = 5)

valueBox(title, value, icon = NULL, color = "blue", width = 5)

infoBox(title, value, icon = NULL, color = "blue", width = 5)
}
\arguments{
\item{title}{Label of the valueBox.}

\item{value}{Value of the valueBox.}

\item{icon}{Icon of the valueBox.}

\item{color}{Color of the valueBox. One of \code{c("", "red", "orange", "yellow", "olive", "green", "teal", "blue", "violet", "purple", "pink", "brown", "grey", "black")}}

\item{width}{Width of the valueBox.}
}
\value{
A valueBox that can be passed to \code{\link[semantic.dashboard]{dashboardBody}}
}
\description{
Create a valueBox with additional UI elements.
}
\section{Functions}{
\itemize{
\item \code{valueBox}: Create a valueBox (alias for \code{value_box} for compatibility with \code{shinydashboard})

\item \code{infoBox}: Create a valueBox (alias for \code{value_box} for compatibility with \code{shinydashboard})
}}

