% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uivalue_box.R
\name{valueBoxes}
\alias{valueBoxes}
\title{Create a group of stacked valueBoxes.}
\usage{
valueBoxes(...)
}
\arguments{
\item{...}{valueBoxes.}
}
\value{
A group of stacked valueBoxes that can be passed to \code{\link[semantic.dashboard]{dashboardBody}}
}
\description{
Create a group of stacked valueBoxes.
}
