% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_dashboard.R
\name{dashboard_header}
\alias{dashboard_header}
\alias{dashboardHeader}
\title{Create a header of a dashboard (alias for \code{dashboard_header} for compatibility with \code{shinydashboard})}
\usage{
dashboard_header(..., color = "black", disable = FALSE)

dashboardHeader(..., color = "black", disable = FALSE)
}
\arguments{
\item{...}{UI elements to include within the header.}

\item{color}{Color of the sidebar. One of \code{c("red", "orange", "yellow", "olive", "green", "teal", "blue", "violet", "purple", "pink", "brown", "grey", "black")}}

\item{disable}{If \code{TRUE}, don't display the header.}
}
\value{
A header that can be passed to \code{\link[semantic.dashboard]{dashboardPage}}
}
\description{
Create a header of a dashboard with other additional UI elements.
}
\section{Functions}{
\itemize{
\item \code{dashboardHeader}: Create a header of a dashboard (alias for \code{dashboard_header} for compatibility with \code{shinydashboard})
}}

