% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_dashboard.R
\name{dashboard_page}
\alias{dashboard_page}
\title{Create a dashboard.}
\usage{
dashboard_page(dashboardSidebar, dashboardBody, title = "",
  suppress_bootstrap = TRUE)
}
\arguments{
\item{dashboardSidebar}{Sidebar of a dashboard.}

\item{dashboardBody}{Body of a dashboard.}

\item{title}{Title of a dashboard.}

\item{suppress_bootstrap}{There are some conflicts in CSS styles between SemanticUI and Bootstrap. For the time being it's better to suppress Bootstrap. If \code{TRUE} bootstrap dependency from \code{shiny} will be disabled.}
}
\value{
Dashboard.
}
\description{
Create a page with menu item sidebar and body containing tabs and other additional elements.
}
