% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_dashboard.R
\name{dashboard_page}
\alias{dashboard_page}
\alias{dashboardPage}
\title{Create a dashboard.}
\usage{
dashboard_page(header, sidebar, body, title = "", suppress_bootstrap = TRUE)

dashboardPage(header, sidebar, body, title = "", suppress_bootstrap = TRUE)
}
\arguments{
\item{header}{Header of a dashboard.}

\item{sidebar}{Sidebar of a dashboard.}

\item{body}{Body of a dashboard.}

\item{title}{Title of a dashboard.}

\item{suppress_bootstrap}{There are some conflicts in CSS styles between SemanticUI and Bootstrap. For the time being it's better to suppress Bootstrap. If \code{TRUE} bootstrap dependency from \code{shiny} will be disabled.}
}
\value{
Dashboard.
}
\description{
Create a page with menu item sidebar and body containing tabs and other additional elements.
}
\section{Functions}{
\itemize{
\item \code{dashboardPage}: Create a dashboard (alias for \code{dashboard_page} for compatibility with \code{shinydashboard})
}}

