% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_dashboard.R
\name{dashboard_sidebar}
\alias{dashboard_sidebar}
\alias{dashboardSidebar}
\title{Create a sidebar of a dashboard.}
\usage{
dashboard_sidebar(..., side = "left", size = "", color = "",
  inverted = FALSE, center = FALSE, visible = TRUE, disable = FALSE)

dashboardSidebar(..., side = "left", size = "", color = "",
  inverted = FALSE, center = FALSE, visible = TRUE, disable = FALSE)
}
\arguments{
\item{...}{UI elements to include within the sidebar.}

\item{side}{Placement of the sidebar. One of \code{c("left", "right", "top", "bottom")}}

\item{size}{Size of the sidebar. One of \code{c("", "thin", "very thin", "wide", "very wide")}}

\item{color}{Color of the sidebar. One of \code{c("red", "orange", "yellow", "olive", "green", "teal", "blue", "violet", "purple", "pink", "brown", "grey", "black")}}

\item{inverted}{Should color be inverted.}

\item{center}{Should label and icon be centerd on menu items. Default to \code{FALSE}}

\item{visible}{Should sidebar be visible on start. Default to \code{TRUE}}

\item{disable}{If \code{TRUE}, don't display the sidebar.}
}
\value{
A sidebar that can be passed to \code{\link[semantic.dashboard]{dashboardPage}}
}
\description{
Create a pushable sidebar of a dashboard with menu items and other additional UI elements.
}
\section{Functions}{
\itemize{
\item \code{dashboardSidebar}: Create a sidebar of a dashboard (alias for \code{dashboard_sidebar} for compatibility with \code{shinydashboard})
}}

