% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_dashboard.R
\name{dashboardBody}
\alias{dashboardBody}
\title{Create a body of a dashboard.}
\usage{
dashboardBody(...)
}
\arguments{
\item{...}{UI elements to include within the body.}
}
\value{
A tab that can be passed to \code{\link[semantic.dashboard]{dashboardPage}}
}
\description{
Create a body of a dashboard with tabs and other additional UI elements.
}
\examples{
if(interactive()){

  library(shiny)
  library(semantic.dashboard)

  ui <- dashboardPage(
    dashboardSidebar(side = "top", size = "thin", color = "teal",
                     uimenu_item("tab1", "Tab 1"),
                     uimenu_item("tab2", "Tab 2")),
    dashboardBody(uitab(id = "tab1", active = TRUE, p("Tab 1")),
                  uitab(id = "tab2", p("Tab 2")))
  )

  server <- function(input, output) {
  }

  shinyApp(ui, server)
}
}
