% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_dashboard.R
\name{dashboardPage}
\alias{dashboardPage}
\title{Create a dashboard.}
\usage{
dashboardPage(dashboardSidebar, dashboardBody, title = "",
  supress_bootsrtap = TRUE)
}
\arguments{
\item{dashboardSidebar}{Sidebar of a dashboard.}

\item{dashboardBody}{Body of a dashboard.}

\item{title}{Title of a dashboard.}

\item{supress_bootsrtap}{If \code{TRUE} bootstrap dependency from \code{shiny} will be disabled.}
}
\value{
Dashboard.
}
\description{
Create a dashboard.
}
\examples{
dashboardPage(
 dashboardSidebar(side = "top",
 uimenu_item("a", "Tab 1"),
 uimenu_item("b", "Tab 2"),
 uimenu_item("c", "Tab 3")),
dashboardBody(uitab(id = "a", active = TRUE,
  numericInput("n1", "n", 50),
  plotOutput("plot1"),
  numericInput("n2", "n", 50),
  plotOutput("plot2"),
  numericInput("n3", "n", 50),
  plotOutput("plot3")),
  uitab(id = "b",
        numericInput("n4", "nn", 50),
        plotOutput("plot4")),
  uitab(id = "c",
        numericInput("n5", "nnn", 50),
        plotOutput("plot5")))
        )
}
