% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_dashboard.R
\name{dashboardPage}
\alias{dashboardPage}
\title{Create a dashboard.}
\usage{
dashboardPage(dashboardSidebar, dashboardBody, title = "",
  suppress_bootsrtap = TRUE)
}
\arguments{
\item{dashboardSidebar}{Sidebar of a dashboard.}

\item{dashboardBody}{Body of a dashboard.}

\item{title}{Title of a dashboard.}

\item{suppress_bootsrtap}{There are some conflicts in CSS styles between SemanticUI and Bootstrap. For the time being it's better to suppress Bootstrap. If \code{TRUE} bootstrap dependency from \code{shiny} will be disabled.}
}
\value{
Dashboard.
}
\description{
Create a page with menu item sidebar and body containing tabs and other additional elements.
}
\examples{
if(interactive()){

  library(shiny)
  library(semantic.dashboard)

  ui <- dashboardPage(
    dashboardSidebar(side = "top", size = "thin", color = "teal",
                     uimenu_item("tab1", "Tab 1"),
                     uimenu_item("tab2", "Tab 2")),
    dashboardBody(uitab(id = "tab1", active = TRUE, p("Tab 1")),
                  uitab(id = "tab2", p("Tab 2")))
  )

  server <- function(input, output) {
  }

  shinyApp(ui, server)
}
}
