% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_dashboard.R
\name{dashboardSidebar}
\alias{dashboardSidebar}
\title{Create a sidebar of a dashboard.}
\usage{
dashboardSidebar(..., side = "left", size = "", color = "black")
}
\arguments{
\item{...}{UI elements to include within the sidebar.}

\item{side}{Placement of the sidebar. One of \code{c("left", "right", "top", "bottom")}}

\item{size}{Size of the sidebar. One of \code{c("", "thin", "very thin", "wide", "very wide")}}

\item{color}{Color of the sidebar.}
}
\value{
A sidebar that can be passed to \code{dashboardPage}
}
\description{
Create a sidebar of a dashboard.
}
\examples{
dashboardSidebar(side = "top",
uimenu_item("a", "Tab 1"),
uimenu_item("b", "Tab 2"),
uimenu_item("c", "Tab 3"))
}
