% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_dashboard.R
\name{dashboardSidebar}
\alias{dashboardSidebar}
\title{Create a sidebar of a dashboard.}
\usage{
dashboardSidebar(..., side = "left", size = "", color = "black")
}
\arguments{
\item{...}{UI elements to include within the sidebar.}

\item{side}{Placement of the sidebar. One of \code{c("left", "right", "top", "bottom")}}

\item{size}{Size of the sidebar. One of \code{c("", "thin", "very thin", "wide", "very wide")}}

\item{color}{Color of the sidebar. One of \code{c("red", "orange", "yellow", "olive", "green", "teal", "blue", "violet", "purple", "pink", "brown", "grey", "black")}}
}
\value{
A sidebar that can be passed to \code{\link[semantic.dashboard]{dashboardPage}}
}
\description{
Create a pushable sidebar of a dashboard with menu items and other additional UI elements.
}
\examples{
if(interactive()){

  library(shiny)
  library(semantic.dashboard)

  ui <- dashboardPage(
    dashboardSidebar(side = "top", size = "thin", color = "teal",
                     uimenu_item("tab1", "Tab 1"),
                     uimenu_item("tab2", "Tab 2")),
    dashboardBody(uitab(id = "tab1", active = TRUE, p("Tab 1")),
                  uitab(id = "tab2", p("Tab 2")))
  )

  server <- function(input, output) {
  }

  shinyApp(ui, server)
}
}
