% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdown_menu.R
\name{message_item}
\alias{message_item}
\alias{messageItem}
\title{Create a message item.}
\usage{
message_item(from, message, ..., icon = "user")

messageItem(from, message, ..., icon = "user")
}
\arguments{
\item{from}{Who the message is from.}

\item{message}{Text of the message.}

\item{...}{Additional UI elements to include within the dropdown menu.}

\item{icon}{Additional icon.}
}
\value{
A message item that can be passed to \code{\link[semantic.dashboard]{dropdownMenu}}
}
\description{
Create a message item.
}
\section{Functions}{
\itemize{
\item \code{messageItem}: Create a message item (alias for \code{message_item} for compatibility with \code{shinydashboard})
}}

\examples{
messageItem("Marek", "Another test!", icon = "warning")
}
