% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu_item.R
\name{sidebar_menu}
\alias{sidebar_menu}
\title{Create a sidebar menu (alias for \code{sidebar_menu} for compatibility with \code{shinydashboard})}
\usage{
sidebar_menu(...)
}
\arguments{
\item{...}{Menu items.}
}
\value{
A sidebar menu that can be passed \code{\link[semantic.dashboard]{dashboardSidebar}}
}
\description{
Create a sidebar menu with menu items.
}
