% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu_item.R
\name{sidebar_menu}
\alias{sidebar_menu}
\alias{sidebarMenu}
\title{Create a sidebar menu.}
\usage{
sidebar_menu(...)

sidebarMenu(...)
}
\arguments{
\item{...}{Menu items.}
}
\value{
A sidebar menu that can be passed \code{\link[semantic.dashboard]{dashboardSidebar}}
}
\description{
Create a sidebar menu with menu items.
}
\section{Functions}{
\itemize{
\item \code{sidebarMenu}: Create a sidebar menu (alias for \code{sidebar_menu} for compatibility with \code{shinydashboard})
}}

\examples{
sidebarMenu(
  menuItem(tabName = "plot_tab", text = "My plot", icon = icon("home")),
  menuItem(tabName = "table_tab", text = "My table", icon = icon("smile"))
  )
}
