% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab_items}
\alias{tab_items}
\alias{tabItems}
\title{Create a panel with tabs.}
\usage{
tab_items(...)

tabItems(...)
}
\arguments{
\item{...}{Tabs.}
}
\value{
A panel with tabs that can be passed to \code{\link[semantic.dashboard]{dashboardBody}}
}
\description{
Create a panel with tabs.
}
\section{Functions}{
\itemize{
\item \code{tabItems}: Create a panel with tabs (alias for \code{tab_items} for compatibility with \code{shinydashboard})
}}

\examples{
tabItems(
 tabItem(tabName = "tab1", "Tab 1"),
 tabItem(tabName = "tab2", "Tab 2"))
}
