% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu_item.R
\name{uimenu_item}
\alias{uimenu_item}
\title{Create a menu item corresponding to a tab.}
\usage{
uimenu_item(id, label)
}
\arguments{
\item{id}{Id of the tab.}

\item{label}{Label of the menu item.}
}
\value{
A menu item that can be passed \code{\link[semantic.dashboard]{dashboardSidebar}}
}
\description{
Create a menu item corresponding to a tab.
}
\examples{
uimenu_item("a", "Tab 1")
}
