% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uitab.R
\name{uitab}
\alias{uitab}
\title{Create a tab panel.}
\usage{
uitab(tabName, ..., active = FALSE)
}
\arguments{
\item{tabName}{Id of the tab.}

\item{...}{UI elements to include within the tab.}

\item{active}{Should it be active.}
}
\value{
A tab that can be passed to \code{\link[semantic.dashboard]{dashboardBody}}
}
\description{
Create a tab panel with additional UI elements.
}
