% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{uitab}
\alias{uitab}
\title{Create a tab panel.}
\usage{
uitab(id, ..., active = FALSE)
}
\arguments{
\item{id}{Id of the tab.}

\item{...}{UI elements to include within the tab.}

\item{active}{Should it be active.}
}
\value{
A tab that can be passed to \code{\link[semantic.dashboard]{dashboardBody}}
}
\description{
Create a tab panel.
}
\examples{
uitab(id = "b",
numericInput("n4", "nn", 50),
plotOutput("plot4"))
}
