% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uivalue_box_output.R
\name{value_box_output}
\alias{value_box_output}
\alias{valueBoxOutput}
\alias{infoBoxOutput}
\title{Create a value box output (alias for \code{value_box_output} for compatibility with \code{shinydashboard})}
\usage{
value_box_output(outputId, width = 5)

valueBoxOutput(outputId, width = 5)

infoBoxOutput(outputId, width = 5)
}
\arguments{
\item{outputId}{Id of the output.}

\item{width}{Width of the valueBox.}
}
\value{
A value box that can be passed to \code{\link[semantic.dashboard]{dashboardBody}}
}
\description{
UI-side function for dynamic valueBox.
}
\section{Functions}{
\itemize{
\item \code{valueBoxOutput}: Create a valueBox output (alias for \code{value_box_output} for compatibility with \code{shinydashboard})

\item \code{infoBoxOutput}: Create a valueBox output (alias for \code{value_box_output} for compatibility with \code{shinydashboard})
}}

